[CS-76] Integrate changes from CS-75

+++ bitprim/idiomatic/INativeList.cs
@@ -6,6 +6,6 @@ namespace Bitprim
 {
     public interface INativeList<T> :  IEnumerable<T>, IDisposable
     {
-        T this[uint index] { get; }
+        T this[UInt64 index] { get; }
     }
 }
